Swap form validation fix
